requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.30.3
httpx==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
itsdangerous==2.2.0
pydantic-settings==2.6.0
sqlmodel==0.0.17
authlib==1.3.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
